requirements.txt
Django==4.2.5
djangorestframework==3.14.0
Pillow==10.0.1
psycopg2-binary==2.9.8
django-filter
djangorestframework-simplejwt
coverage
stripe
python-dotenv
drf-yasg
